requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.9.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
